#33 Writing stub modules and methods
